% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3calendar.R
\name{d3calendar}
\alias{d3calendar}
\title{Calendar Heatmap}
\usage{
d3calendar(df, width = NULL, height = NULL, elementId = NULL,
  settings = list(n_colors = 11))
}
\arguments{
\item{df}{A data.frame object, which should has at least two columns: "date" and "value".}
}
\description{
Plot a calendar heatmap base on D3.
}
